+++ source/games/duke/src/input.cpp
@@ -369,7 +369,7 @@ void hud_input(int plnum)
 					{
 						int j = gs.max_player_health - p->GetActor()->s->extra;
 
-						if ((unsigned int)p->firstaid_amount > j)
+						if (p->firstaid_amount > j)
 						{
 							p->firstaid_amount -= j;
 							p->GetActor()->s->extra = gs.max_player_health;
@@ -568,13 +568,13 @@ static double motoApplyTurn(player_struct* p, ControlInfo* const hidInput, bool 
 
 		if (kbdLeft || hidInput->mouseturnx < 0 || hidInput->dyaw < 0)
 		{
-			p->TiltStatus -= factor;
+			p->TiltStatus -= (float)factor;
 			if (p->TiltStatus < -10)
 				p->TiltStatus = -10;
 		}
 		else if (kbdRight || hidInput->mouseturnx > 0 || hidInput->dyaw > 0)
 		{
-			p->TiltStatus += factor;
+			p->TiltStatus += (float)factor;
 			if (p->TiltStatus > 10)
 				p->TiltStatus = 10;
 		}
@@ -589,7 +589,7 @@ static double motoApplyTurn(player_struct* p, ControlInfo* const hidInput, bool 
 			if (kbdLeft || p->moto_drink < 0 || hidInput->mouseturnx < 0 || hidInput->dyaw < 0)
 			{
 				updateTurnHeldAmt(factor);
-				p->TiltStatus -= factor;
+				p->TiltStatus -= (float)factor;
 
 				if (p->TiltStatus < -10)
 					p->TiltStatus = -10;
@@ -607,7 +607,7 @@ static double motoApplyTurn(player_struct* p, ControlInfo* const hidInput, bool 
 			if (kbdRight || p->moto_drink > 0 || hidInput->mouseturnx > 0 || hidInput->dyaw > 0)
 			{
 				updateTurnHeldAmt(factor);
-				p->TiltStatus += factor;
+				p->TiltStatus += (float)factor;
 
 				if (p->TiltStatus > 10)
 					p->TiltStatus = 10;
@@ -627,9 +627,9 @@ static double motoApplyTurn(player_struct* p, ControlInfo* const hidInput, bool 
 			resetTurnHeldAmt();
 
 			if (p->TiltStatus > 0)
-				p->TiltStatus -= factor;
+				p->TiltStatus -= (float)factor;
 			else if (p->TiltStatus < 0)
-				p->TiltStatus += factor;
+				p->TiltStatus += (float)factor;
 		}
 	}
 
@@ -663,7 +663,7 @@ static double boatApplyTurn(player_struct *p, ControlInfo* const hidInput, bool 
 
 				if (!p->NotOnWater)
 				{
-					p->TiltStatus -= factor;
+					p->TiltStatus -= (float)factor;
 					if (p->TiltStatus < -10)
 						p->TiltStatus = -10;
 				}
@@ -684,7 +684,7 @@ static double boatApplyTurn(player_struct *p, ControlInfo* const hidInput, bool 
 
 				if (!p->NotOnWater)
 				{
-					p->TiltStatus += factor;
+					p->TiltStatus += (float)factor;
 					if (p->TiltStatus > 10)
 						p->TiltStatus = 10;
 				}
@@ -704,9 +704,9 @@ static double boatApplyTurn(player_struct *p, ControlInfo* const hidInput, bool 
 			resetTurnHeldAmt();
 
 			if (p->TiltStatus > 0)
-				p->TiltStatus -= factor;
+				p->TiltStatus -= (float)factor;
 			else if (p->TiltStatus < 0)
-				p->TiltStatus += factor;
+				p->TiltStatus += (float)factor;
 		}
 	}
 	else if (!p->NotOnWater)
@@ -714,9 +714,9 @@ static double boatApplyTurn(player_struct *p, ControlInfo* const hidInput, bool 
 		resetTurnHeldAmt();
 
 		if (p->TiltStatus > 0)
-			p->TiltStatus -= factor;
+			p->TiltStatus -= (float)factor;
 		else if (p->TiltStatus < 0)
-			p->TiltStatus += factor;
+			p->TiltStatus += (float)factor;
 	}
 
 	if (fabs(p->TiltStatus) < factor)
@@ -747,12 +747,12 @@ static void processVehicleInput(player_struct *p, ControlInfo* const hidInput, I
 
 	if (p->OnMotorcycle)
 	{
-		input.avel = motoApplyTurn(p, hidInput, kbdLeft, kbdRight, scaleAdjust);
+		input.avel = (float)motoApplyTurn(p, hidInput, kbdLeft, kbdRight, scaleAdjust);
 		if (p->moto_underwater) p->MotoSpeed = 0;
 	}
 	else
 	{
-		input.avel = boatApplyTurn(p, hidInput, kbdLeft, kbdRight, scaleAdjust);
+		input.avel = (float)boatApplyTurn(p, hidInput, kbdLeft, kbdRight, scaleAdjust);
 	}
 
 	loc.fvel = clamp(xs_CRoundToInt(p->MotoSpeed), -(MAXVELMOTO >> 3), MAXVELMOTO);
